requirements.txt
Flask==2.0.2
paho-mqtt==1.6.1
werkzeug==2.0.3

